src/utils/notebook_requirements.txt
datalib-n26[postgres]==0.14.1
jupyter
jupyterlab